Calculadora-de-Idade/requirements.txt
Babel==2.11.0
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
tkcalendar==1.6.1
